Upload of firmwares
BEWARE:
- `crkbd_ecwl_default.hex` is compiled with gonnoc's own fork and branch, based on the old QMK 6.1
-  `crkbd_ecrevival_default.hex` is compiled with the current stable QMK, it has some flaws, described below, but the general functionality is preserved.

As of now the Original ECWL has a wider range of sensibility. The EC revival one doesn't have that fine control. I'm working on a patch/rework of the custom matrix implementation to solve this.

For proper testing and fine tuning I STRONGLY advice to self compile the branch that gonnoc has (qmk_firmware, dev/ecwl).